Moved common functions for getting the network info to a separate module
